feat:move downlode crypto files from shuffle.ts to utility.ts
