+++ packages/jssdk/package.json
@@ -7,7 +7,7 @@
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.2.1",
     "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@poseidon-zkp/poseidon-zk-proof": "0.0.1",
+    "@poseidon-zkp/poseidon-zk-proof": "0.0.6",
     "@typechain/ethers-v5": "^10.1.1",
     "@typechain/hardhat": "^6.1.4",
     "@types/chai": "^4.3.4",
